// frontend/src/lib/protocol/generated/NavigationTarget.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type NavigationTarget = { "type": "traffic", id: string, } | { "type": "mapPosition", latitudeDegrees: number, longitudeDegrees: number, } | { "type": "waypoint", name: string, latitudeDegrees: number, longitudeDegrees: number, elevationMeters: number, };